Add implementations of 'safe' functions to solve error handling
The fpinscala book did not cover how to safely handle errors that
occur when trying to perform computations in Par code.

I updated versions of Future, Par, and the functions that allow
for computation to communicate failure via the Option type. The
way it works is that you use the 'safe' basis functions (unit,
lazyUnit, fork, map, map2, etc...) to build a SafePar. Then,
safeRun is called with it and will get a None in the presence of
a failure. Could use an either, but this was mainly to capture an
example that at least indicates that something went wrong through
None.