Checkpoint commit for Chapter 8 Prop code
Rearranged some of the code here because I brought in a bunch of
the code that I had used previously just with scala/scalac compiling
instead of sbt. I wanted to use some of the tools created in the
previous chapter, which just meant organizing these other files into
the sbt project.