Add runAll and oneOf Prop methods and test some more properties
I was able to play around a little to get a better richPar generator.
This involved creating the oneOf function to allow me to build a
generator that would take a list of options and choose between them
randomly. I used this to generate Par.{unit, lazyUnit} calls and
to choose a binary relation for reducing a list of random pars.

I used the generator that was created from this effort to then create
a property that gave more variance to the 8.16 exercise work that had
previously been done.

runAll allows an easier specification to run a list of properties in
sequence.