Get the underlying parse functions working
Chapter 9 was a bit of a bear. I had to consult the github repo and then come up
with a new strategy, the one the book presented tried to do too much without
testing to see that anything was working.

The code I have in this commit allows for a base parser implementation to be
experimented with for testing the core base api functionality. This includes
the unit, string, flatMap, map2, map, many, and many1 functions. Small examples
were added to the ParseRepl to demonstrate how to use them.

Worth noting here that Parsers [Parser[+_]] as a higher kinded type is an important
thing to grok. In the base parser, I gave Parser[+_] a type definition of a function
that takes a ParseState and returns a Result. This is just one possible interpretation,
another class could implement it in whatever way they choose. This is powerful, but
also confusing this first time around because the syntax is a bit mysterious as to
how the trait declaration necessitates a programmer defined implementation when
trying to create a class that uses it.

I found the paper [Generics of a Higher Kind](http://adriaanm.github.io/files/higher.pdf) to be helpful.

In order to play with this code, the following changes were made to the Prop lib:

Updated SGen to include apply and **
Updated Gen to include string and stringN
Update Prop to include equal